Extract shared and add test.
